small refactor to move grouped select to a separate file
