+++ agenta-web/src/components/Playground/Views/ParametersCards.tsx
@@ -3,17 +3,6 @@ import {createUseStyles} from "react-jss"
 import {renameVariables} from "@/lib/helpers/utils"
 import {Parameter, InputParameter} from "@/lib/Types"
 import {DownOutlined} from "@ant-design/icons"
-import {
-    IconType,
-    OpenAI,
-    Mistral,
-    Cohere,
-    Anthropic,
-    Perplexity,
-    Together,
-    OpenRouter,
-    Fireworks,
-} from "@lobehub/icons"
 
 import {
     Row,
@@ -29,6 +18,7 @@ import {
     Menu,
     Space,
 } from "antd"
+import { GroupedSelect } from "./GroupedSelect"
 
 const useStyles = createUseStyles({
     row1: {
@@ -83,80 +73,8 @@ const useStyles = createUseStyles({
     },
     select: {
         width: "100%",
-    },
-    option: {
-        display: "flex",
-        alignItems: "center",
-        gap: "8px",
-    },
-})
-
-interface IconMap {
-    [key: string]: IconType
-}
-
-interface GroupedSelectProps {
-    choices: {[group: string]: string[]}
-    defaultValue: string
-    handleChange: (value: string) => void
-}
-
-const iconMap: {[key: string]: React.ComponentType<any>} = {
-    "Open AI": OpenAI,
-    "Mistral AI": Mistral.Color,
-    Cohere: Cohere.Color,
-    Anthropic: Anthropic,
-    "Perplexity AI": Perplexity.Color,
-    "Together AI": Together.Color,
-    OpenRouter: OpenRouter,
-}
-
-const getTextContent = (element: React.ReactNode): string => {
-    if (typeof element === "string") {
-        return element
-    } else if (React.isValidElement(element) && element.props.children) {
-        return React.Children.toArray(element.props.children).reduce<string>(
-            (acc, child) => acc + getTextContent(child),
-            "",
-        )
     }
-    return ""
-}
-
-const filterOption = (input: string, option?: {label: React.ReactNode; value: string}) =>
-    getTextContent(option?.label).toLowerCase().includes(input.toLowerCase())
-
-const GroupedSelect: React.FC<GroupedSelectProps> = ({choices, defaultValue, handleChange}) => {
-    const classes = useStyles()
-
-    const options = Object.entries(choices).map(([groupLabel, groupChoices]) => ({
-        label: (
-            <div className={classes.option}>
-                {iconMap[groupLabel] ? React.createElement(iconMap[groupLabel]) : null}
-                {groupLabel}
-            </div>
-        ),
-        options: groupChoices.map((choice) => ({
-            label: (
-                <div className={classes.option}>
-                    {iconMap[groupLabel] ? React.createElement(iconMap[groupLabel]) : null}
-                    {choice}
-                </div>
-            ),
-            value: choice,
-        })),
-    }))
-    return (
-        <Select
-            showSearch
-            defaultValue={defaultValue}
-            className={classes.select}
-            onChange={handleChange}
-            filterOption={filterOption}
-            options={options as any}
-        />
-    )
-}
+})
 
 interface ModelParametersProps {
     optParams: Parameter[] | null
